iv_handle: Don't lose HANDLE active status when setting handler to NULL.
When the handler of an active iv_handle is set to NULL and then back
again to a non-NULL value, and the underlying HANDLE is of the auto
resetting variety, we would fail to invoke the newly installed handler
in the next poll cycle, and the active status of the underlying HANDLE
would effectively be lost.

Fix this by putting iv_handle objects that are active but currently
have no associated handler on a separate active list, and moving them
to the main active list when their handler is set to a non-NULL value.
(And skip calling WaitForSingleObjectEx() in iv_handle_poll_and_run()
if there are active HANDLEs when we enter the function.)

Signed-off-by: Lennert Buytenhek <buytenh@wantstofly.org>